ri-etu.dtb: add expansion header GPIO as LED
This is for use with the golfball machine that needs to assert
a 300 ms signal to trigger the feeder.  The transient LED trigger
allows us to just trigger such a signal.

Fits nicely with a udev rule such as:

SUBSYSTEM=="leds",KERNEL=="extswitch",TAG+="poc",ENV{POC_CODRIVER}="cow",ATTR{duration}="200",ATTR{state}="1"