Add SPI1 for fiscal memory
The SPI1 bus terminates on an FPC connector.  The idea is to connect the
"fiscal memory" there.  We'll need a driver for the fiscal memory once
we know how it's supposed to work; as such, we define the SPI bus
currently but don't add the actual "device" until a driver for it
exists.