Add 'strobe' LED
The CPU_STROBE signal is a signal to the MCU that indicates that the CPU
is running.  This signal just needs to be toggled regularly, which is
effectively what the 'timer' LED trigger does.  As such, we set up the
STROBE output as a 'virtual LED' and let the timer trigger take care of
toggling it for us.